bot/requirements.txt
aiogram==3.20.0
aiohttp==3.13.3
supabase==2.15.1
python-dotenv==1.1.0
